Move Guardian services to entity platform services
